pipeline-tasks/requirements.txt
fastapi
fastapi[standard]
uvicorn
jinja2
python-multipart
azure-functions
azure-keyvault-secrets
azure.identity